[src/charm.py] Add charm-resource gitea-binary to charm.py
See LP Bug: https://pad.lv/2069052
See LP Bug: https://pad.lv/2069056

The following updates allow us to supply a charm resource when deploying
gitea. The resource is in fact the binary itself. See bug for more
information.

In addition to adding support for a charm resource I have refactored the
_on_install(self, event). Specifically moving the code responsible for
creating all of the necessary directories to its own method.  This will
allow the charm to halt the installation in the event there is an error.

The following changes have been tested against a locally bootstrapped
controller.

Signed-off-by: Sean Feole <sean.feole@canonical.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Signed-off-by: Sean Feole <sean.feole@canonical.com>